Provide `saveEntries` and `loadEntries` methods via injected `StorageService`.
